obfucated sensitive strings, extracted to .env file
