+++ .gitignore
@@ -1,3 +1,7 @@
+./src/config.js
+
+
+
 # OSX
 #
 .DS_Store
