+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.BC",
+    package = "com.superappidea",
 )
 
 android_resource(
     name = "res",
-    package = "com.BC",
+    package = "com.superappidea",
     res = "src/main/res",
 )
