+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">BC</string>
+    <string name="app_name">SuperAppIdea</string>
 </resources>
