+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "BC",
-  "displayName": "BC"
+  "name": "SuperAppIdea",
+  "displayName": "SuperAppIdea"
 }
