+++ ios/Podfile
@@ -3,7 +3,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '11.0'
 
-target 'BC' do
+target 'SuperAppIdea' do
   config = use_native_modules!
 
   use_react_native!(
@@ -12,10 +12,6 @@ target 'BC' do
     :hermes_enabled => false
   )
 
-  target 'BCTests' do
-    inherit! :complete
-    # Pods for testing
-  end
 
   # Enables Flipper.
   #
