+++ ios/SuperAppIdea.xcodeproj/xcshareddata/xcschemes/SuperAppIdea.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "BC.app"
-               BlueprintName = "BC"
-               ReferencedContainer = "container:BC.xcodeproj">
+               BuildableName = "SuperAppIdea.app"
+               BlueprintName = "SuperAppIdea"
+               ReferencedContainer = "container:SuperAppIdea.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "BCTests.xctest"
-               BlueprintName = "BCTests"
-               ReferencedContainer = "container:BC.xcodeproj">
+               BuildableName = "SuperAppIdeaTests.xctest"
+               BlueprintName = "SuperAppIdeaTests"
+               ReferencedContainer = "container:SuperAppIdea.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BC.app"
-            BlueprintName = "BC"
-            ReferencedContainer = "container:BC.xcodeproj">
+            BuildableName = "SuperAppIdea.app"
+            BlueprintName = "SuperAppIdea"
+            ReferencedContainer = "container:SuperAppIdea.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BC.app"
-            BlueprintName = "BC"
-            ReferencedContainer = "container:BC.xcodeproj">
+            BuildableName = "SuperAppIdea.app"
+            BlueprintName = "SuperAppIdea"
+            ReferencedContainer = "container:SuperAppIdea.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
