+++ ios/SuperAppIdea/Images.xcassets/AppIcon.appiconset/Contents.json
