+++ ios/SuperAppIdeaTests/SuperAppIdeaTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface BCTests : XCTestCase
+@interface SuperAppIdeaTests : XCTestCase
 
 @end
 
-@implementation BCTests
+@implementation SuperAppIdeaTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
