+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "BC",
+  "name": "superappidea",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -9,9 +9,6 @@
     "test": "jest",
     "lint": "eslint ."
   },
-  "jest": {
-    "preset": "react-native"
-  },
   "dependencies": {
     "@react-native-community/masked-view": "^0.1.11",
     "@react-navigation/native": "^6.0.2",
@@ -20,6 +17,7 @@
     "react": "17.0.2",
     "react-native": "0.65.1",
     "react-native-chart-kit": "^6.11.0",
+    "react-native-codegen": "^0.0.8",
     "react-native-gesture-handler": "^1.10.3",
     "react-native-modal": "^13.0.0",
     "react-native-reanimated": "^2.2.1",
@@ -35,15 +33,16 @@
     "saga": "^4.0.0-alpha"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.5",
-    "@babel/preset-env": "^7.15.6",
+    "@babel/core": "^7.12.9",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
     "babel-jest": "^26.6.3",
     "eslint": "7.14.0",
     "jest": "^26.6.3",
-    "metro-react-native-babel-preset": "^0.66.0",
-    "react-native-codegen": "^0.0.7",
+    "metro-react-native-babel-preset": "^0.66.2",
     "react-test-renderer": "17.0.2"
+  },
+  "jest": {
+    "preset": "react-native"
   }
 }
